RHOAIENG-79883: Move uriToModelLocation from frontend to @odh-dashboard/k8s-core
