+++ packages/model-serving/modelRegistry/PreWizardDeployModal.tsx
@@ -15,12 +15,11 @@ import {
 import { Link } from 'react-router';
 import { typedObjectImage, ProjectObjectType } from '@odh-dashboard/ui-core';
 import type { ProjectKind, Connection } from '@odh-dashboard/k8s-core';
-import { getConnectionTypeRef } from '@odh-dashboard/k8s-core';
+import { getConnectionTypeRef, uriToModelLocation } from '@odh-dashboard/k8s-core';
 import { ProjectsContext } from '@odh-dashboard/ui-core/context/ProjectsContext';
 import ProjectSelector from '@odh-dashboard/internal/pages/modelServing/screens/projects/InferenceServiceModal/ProjectSelector';
 import useServingConnections from '@odh-dashboard/internal/pages/projects/screens/detail/connections/useServingConnections';
 import { useWatchConnectionTypes } from '@odh-dashboard/internal/utilities/useWatchConnectionTypes';
-import { uriToModelLocation } from '@odh-dashboard/internal/concepts/modelRegistry/utils';
 import type { ModelDeployPrefillInfo } from '@odh-dashboard/model-registry/shared';
 import useRegistryConnections from './useRegistryConnections';
 import { useExtractFormDataFromRegistry } from './useExtractFormDataFromRegistry';
