+++ packages/model-serving/modelRegistry/useExtractFormDataFromRegistry.ts
@@ -7,9 +7,9 @@ import {
   INFERENCE_SERVICE_NAME_INVALID_CHARS_MESSAGE,
   INFERENCE_SERVICE_NAME_REGEX,
   AccessTypes,
+  uriToModelLocation,
 } from '@odh-dashboard/k8s-core';
 import type { ConnectionTypeValueType } from '@odh-dashboard/k8s-core';
-import { uriToModelLocation } from '@odh-dashboard/internal/concepts/modelRegistry/utils';
 import { useWatchConnectionTypes } from '@odh-dashboard/internal/utilities/useWatchConnectionTypes';
 import type { ModelDeployPrefillInfo } from '@odh-dashboard/model-registry/shared';
 import { getModelRegistryMetadata } from './utils/deployUtils';
